requirements.txt
Django==3.2.9
djangorestframework==3.12.4
mysqlclient==2.0.1
django-mysql==2.2.0
django-storages==1.12.2
boto3==1.20.0
gunicorn==20.1.0
flake8
